+++ test/unit/fixtures/config/config.add-plugins.js
@@ -1,0 +1,3 @@
+module.exports = {
+  addPlugins: ["foo2"]
+};