+++ test/unit/fixtures/config/config.function.js
@@ -1,0 +1,7 @@
+module.exports = () => ({
+  plugins: ["foo"],
+  options: {
+    log: "silly",
+    delay: 1000
+  }
+});