+++ test/unit/fixtures/config/config.modify.js
@@ -1,0 +1,7 @@
+module.exports = config => {
+  delete config.disableCommandLineArguments;
+  config.plugins.push("foo2");
+  delete config.options.foo;
+  config.options.foo3 = "foo3";
+  return config;
+};